ver 0.9.1 	modified:   DockerInitFiles.sh 	modified:   README.md 	modified:   config/version 	modified:   database/db.go 	modified:   database/model/model.go 	modified:   go.mod 	modified:   go.sum 	modified:   install.sh 	modified:   media/inbounds-dark.png 	modified:   media/inbounds.png 	modified:   sub/subService.go 	modified:   web/assets/css/custom.css 	modified:   web/assets/js/langs.js 	modified:   web/assets/js/model/dbinbound.js 	deleted:    web/assets/js/model/models.js 	modified:   web/assets/js/model/setting.js 	modified:   web/assets/js/model/xray.js 	modified:   web/assets/js/util/common.js 	modified:   web/assets/js/util/utils.js 	modified:   web/controller/api.go 	modified:   web/controller/inbound.go 	modified:   web/controller/setting.go 	modified:   web/controller/xraySetting.go 	modified:   web/entity/entity.go 	modified:   web/html/common/js.html 	modified:   web/html/common/qrcode_modal.html 	modified:   web/html/common/text_modal.html 	modified:   web/html/login.html 	modified:   web/html/xui/client_bulk_modal.html 	modified:   web/html/xui/client_modal.html 	modified:   web/html/xui/component/setting.html 	modified:   web/html/xui/form/client.html 	modified:   web/html/xui/form/inbound.html 	modified:   web/html/xui/form/protocol/trojan.html 	modified:   web/html/xui/form/protocol/vless.html 	modified:   web/html/xui/form/stream/stream_settings.html 	modified:   web/html/xui/form/stream/stream_tcp.html 	modified:   web/html/xui/form/tls_settings.html 	modified:   web/html/xui/inbound_client_table.html 	modified:   web/html/xui/inbound_info_modal.html 	modified:   web/html/xui/inbound_modal.html 	modified:   web/html/xui/inbounds.html 	modified:   web/html/xui/index.html 	modified:   web/html/xui/settings.html 	modified:   web/html/xui/xray.html 	modified:   web/service/config.json 	modified:   web/service/inbound.go 	modified:   web/service/server.go 	modified:   web/service/setting.go 	modified:   web/service/tgbot.go 	modified:   web/service/xray.go 	modified:   web/translation/translate.en_US.toml 	modified:   web/translation/translate.fa_IR.toml 	modified:   web/translation/translate.ru_RU.toml 	modified:   web/translation/translate.zh_Hans.toml 	modified:   web/web.go 	modified:   x-ui.sh 	modified:   xray/inbound.go
	media/outbounds.png
	media/rules.png
	web/assets/codemirror/
	web/assets/js/model/outbound.js
	web/html/xui/form/outbound.html
	web/html/xui/form/stream/external_proxy.html
	web/html/xui/xray_outbound_modal.html
	web/html/xui/xray_reverse_modal.html
	web/html/xui/xray_rule_modal.html
	web/translation/translate.vi_VN.toml

+++ install.sh
@@ -99,7 +99,17 @@ else
 fi
 sleep 3
 if [[ ! -f "/etc/letsencrypt/live/${SubDomain}.${MainDomain}/privkey.pem" ]]; then
-	msg_err "$SubDomai}.$MainDomain SSL certificate could not be generated, Maybe the domain or IP domain is invalid!" && exit 1
+	msg_err "$SubDomain.$MainDomain SSL certificate could not be generated, Maybe the domain or IP domain is invalid!" && exit 1
+fi
+
+if [[ ! -f "/etc/letsencrypt/live/$sub.${MainDomain}/privkey.pem" ]]; then
+	certbot certonly --standalone --non-interactive --force-renewal --agree-tos --register-unsafely-without-email --cert-name "sub.$MainDomain" -d "$domain"
+else
+	msg_ok "sub.$MainDomain SSL Certificate is exist!"
+fi
+sleep 3
+if [[ ! -f "/etc/letsencrypt/live/sub.${MainDomain}/privkey.pem" ]]; then
+	msg_err "sub.$MainDomain SSL certificate could not be generated, Maybe the domain or IP domain is invalid!" && exit 1
 fi
 
 cat > "/etc/nginx/sites-available/$MainDomain" << EOF
@@ -199,6 +209,7 @@ listen front
  use_backend CDN if { req.ssl_sni -m end $SubDomain.$MainDomain }
  use_backend REALITY_TROJAN if { req.ssl_sni -m end trojan.$MainDomain }
  use_backend REALITY_VLESS if { req.ssl_sni -m end vless.$MainDomain }
+ use_backend SUBS if { req.ssl_sni -m end sub.$MainDomain }
 
 
 backend CDN
@@ -212,6 +223,10 @@ backend REALITY_VLESS
 backend REALITY_TROJAN
  mode tcp
  server srv1 127.0.0.1:23443
+ 
+backend SUBS
+ mode tcp
+ server srv1 127.0.0.1:2096
 EOF
 ###################################Enable Site###############################
 if [[ -f "/etc/nginx/sites-available/$MainDomain" ]]; then
@@ -276,11 +291,12 @@ if systemctl is-active --quiet x-ui && [[ $XUIPORT -eq $PORT ]]; then clear
 	nginx -T | grep -i 'ssl_certificate\|ssl_certificate_key'
 	msg_inf "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
 	certbot certificates | grep -i 'Path:\|Domains:\|Expiry Date:'
-	msg_inf "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
-	msg_inf "\nTo use the VLESS+reality, create the vless.${MainDomain} in DNS-only mode        "
-	msg_inf "To use the TROJAN+reality, create the trojan.${MainDomain} in DNS-only mode        "
-	msg_inf "\nX-UI Admin Panel: https://${domain}/${RNDSTR}/\n"
-	msg_inf "  Login  | Password"
+	msg_inf "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -\n"
+	msg_inf "To use the VLESS+reality,  create the vless.${MainDomain} in DNS-only mode     "
+	msg_inf "To use the TROJAN+reality, create the trojan.${MainDomain} in DNS-only mode    "
+	msg_inf "But you can change it with command 'nano /etc/haproxy/haproxy.cfg'             \n"
+	msg_inf "\nEX-UI Admin Panel: https://${domain}/${RNDSTR}/\n"
+	msg_inf " Login | Password"
 	sqlite3 -batch $XUIDB 'SELECT "username","password" FROM users;'
 	msg_inf "- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -"
 else
